requirements.txt
aiogram==3.*
aiosqlite
apscheduler
pytz
python-dateutil
